hotfix/8991 rename many files, fix quests component
