+++ components/app/pages/common/userInfoOther.vue
@@ -5,16 +5,9 @@
     <div class="col info-grid__col_left">
       <div class="info-grid__avatar">
         <img
-          v-if="imageData"
           class="info-grid__avatar"
-          :src="imageData"
-          :alt="localUserData.firstName"
-        >
-        <img
-          v-else-if="!imageData"
-          class="info-grid__avatar"
-          src="~/assets/img/app/avatar_empty.png"
-          :alt="localUserData.firstName"
+          src="~/assets/img/temp/avatar.jpg"
+          alt=""
         >
       </div>
       <div class="rating" />
@@ -26,79 +19,59 @@
       </nuxt-link>
     </div>
     <div class="col info-grid__col">
-      <div
-        v-if="firstName && lastName"
-        class="title"
-      >
-        {{ firstName }} {{ lastName }}
+      <div class="title">
+        {{ name }}
       </div>
       <div
         v-if="userRole === 'employer'"
         class="subtitle"
       >
-        {{ company || userInfo.company }}
+        {{ company }}
       </div>
       <div
-        v-if="userDesc"
+        v-if="userRole === 'employer'"
         class="description"
       >
         {{ userDesc }}
       </div>
       <div v-if="selected === 1 && userRole === 'worker' ">
-        <div
-          v-if="userEducations.length > 0"
-        >
+        <div class="knowledge__text">
+          {{ $t('profile.educations') }}
+        </div>
+        <div class="work-exp__container">
           <div
-            class="knowledge__text"
+            class="work-exp__item"
           >
-            {{ $t('profile.educations') }}
-          </div>
-          <div
-            v-if="userEducations"
-            class="work-exp__container"
-          >
-            <div
-              v-for="(item, i) in userEducations"
-              :key="i"
-              class="work-exp__item"
-            >
-              <span class="work-exp__company">{{ item.place }}</span>
-              <span class="work-exp__term">{{ item.from }} - {{ item.to }}</span>
-            </div>
+            <span class="work-exp__company">{{ userEducations.place }}</span>
+            <span class="work-exp__term">{{ userEducations.to }} - {{ userEducations.from }}</span>
           </div>
         </div>
-        <div
-          v-if="userWorkExp.length > 0"
-        >
+        <div class="work-exp__text">
+          {{ $t('profile.prevWorkExp') }}
+        </div>
+        <div class="work-exp__container">
           <div
-            class="work-exp__text"
+            v-for="(item, i) in userWorkExp"
+            :key="i"
+            class="work-exp__item"
           >
-            {{ $t('profile.prevWorkExp') }}
-          </div>
-          <div class="work-exp__container">
-            <div
-              v-for="(item, i) in userWorkExp"
-              :key="i"
-              class="work-exp__item"
-            >
-              <span class="work-exp__company">{{ item.place }}</span>
-              <span class="work-exp__term">{{ item.from }} - {{ item.to }}</span>
-            </div>
+            <span class="work-exp__company">{{ item.place }}</span>
+            <span class="work-exp__term">{{ item.to }} - {{ item.from }}</span>
           </div>
         </div>
       </div>
       <!-- socials links -->
       <div class="socials">
-        <SocialPanel />
+        <socialPanel />
       </div>
       <!-- contacts -->
       <div class="contacts__grid">
         <div class="contacts">
-          <ContactPanel />
+          <contactPanel />
           <div class="btn__container">
             <base-btn
               v-if="userRole === 'worker'"
-              @click="toRaisedViews()"
+              @click="toRaisedViews"
             >
               {{ $t('profile.raiseViews') }}
             </base-btn>
@@ -126,18 +99,19 @@
         </div>
       </div>
     </div>
+
     <div class="share-btn" />
   </div>
 </template>
 
 <script>
 import { mapGetters } from 'vuex';
-import ContactPanel from '~/components/app/Panels/Contact';
-import SocialPanel from '~/components/app/Panels/Social';
+import contactPanel from '~/components/app/panels/contactOther';
+import socialPanel from '~/components/app/panels/social';
 
 export default {
   name: 'UserInfo',
-  components: { ContactPanel, SocialPanel },
+  components: { contactPanel, socialPanel },
   props: {
     selected: {
       type: Number,
@@ -154,15 +128,14 @@ export default {
       tags: 'ui/getTags',
       userRole: 'user/getUserRole',
       userData: 'user/getUserData',
-      userInfo: 'data/getUserInfo',
       imageData: 'user/getImageData',
-      firstName: 'user/getFirstName',
-      lastName: 'user/getLastName',
-      company: 'user/getUserCompany',
-      userDesc: 'user/getUserDesc',
-      userEducations: 'user/getUserEducations',
-      userWorkExp: 'user/getUserWorkExp',
+      userEducations: 'data/getEducations',
+      userWorkExp: 'data/getWorkExp',
+      userInfo: 'data/getUserInfo',
       quest: 'data/getQuest',
+      name: 'data/getName',
+      company: 'data/getCompany',
+      userDesc: 'data/getUserDesc',
     }),
   },
   async mounted() {
@@ -297,6 +270,7 @@ export default {
 
   .avatar {
     border-radius: 89px;
+    object-fit: cover;
   }
 
   .rating {
@@ -348,7 +322,6 @@ export default {
     width: 142px;
     height: 142px;
     border-radius: 50%;
-    object-fit: cover;
   }
   &__col {
     &_left {
@@ -372,8 +345,6 @@ export default {
     align-items: center;
     .col {
       margin-bottom: 10px;
-      margin-left: auto;
-      margin-right: auto;
     }
   }
 }
